Search/Gitlab/requirements.txt
python-gitlab
tenacity